search-jobs-on-xls/pyproject.toml
[tool.poetry]
name = "search-jobs-on-xls"
version = "0.1.0"
description = ""
authors = ["José Pinheiro <josepsousa2012@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
